Update messages.json
Fixing replace post to tweet function to not conflict with portuguese words. Words like reply (Resposta) contain "post" inside it and turn into ("Restweeta") which is not a word at all.

Inserted a prerequisite to have a " " before the word post for it to be replaced. But there should be a way to change the color of a word when it is replaced, because it does conflict with other words. If someone is posting something at the mail, and tweets about it, it would translate to them tweeting somethng at the mail... The function is useful but a change of color would be useful.

Same changes apply to replace repost>retweet